queue: generalize size
Allow arbitrary mapping from queue item to item size. Currently this
mapping is fixed, each item counts as 1 towards the max size. This
doesn't allow for maintaining a correct max size when the items have
wildy varying contribution towards the max size.

The intent behind this change is to allow the original purpose of the
max size to be achived with any kind of `T` (item type): limit the
amount of memory that the queue can consume. The current system only
works for those `T` types that have a fixed memory consumption, or a
non-fixed one with minimal variance. In this case one can set the max
size such that the total memory consumption of the queue stays below
the desired limit (L). Given the memory consumption of an item (M), one
can set max size to L/M to achieve this. But as mentioned this will not
work if the items have wildly different memory consumption. To enable
this case as well, add a `Measurer` template parameter to `queue`. This
is a functor that allows arbitrary mappings from `T` to `size_t`. This
functor is now called for every item added/removed from the queue.

`Measurer` is defaulted to `measure_as_one` which maps all items to 1,
effectively preserving the previous behaviour, so existing code can
continue to use `queue` as before. It is worth to mention that due to
the refactoring of the code, the queue will now take one `size_t` more
memory than before. I think this is acceptable given the benefits of
being able to use the queue for any `T` now.
Another consequence of the change is that `size()` now returns the
current size of the queue as measured by the measurer, not the number of
items in the queue (as before). For that, a new method: `item_count()`
is added. When using the default `Measurer` the two will be the same.