queue: generalize size
Allow arbitrary types to be used as max size and arbitrary mappings from
queue items to size. To this end, this patch introduces a new template
parameter for queue: `LimitPolicy`. This policy is responsible for
determining when the queue is full. In addition to the already existent
`_max`, the queue now also maintains an `_actual` member which contains
the current state of the queue's size. `_actual` is managed entirely by
the limit policy. When adding or removing an item, `_actual`, as well as
the to be added/removed item is passed to the limit policy, which should
increment/decrement it as appropriate. The type of `_max` and `_actual`
is `LimitPolicy::size_type`.

`LimitPolicy` is defaulted to `ItemCountLimitPolicy` which limits the
size of the queue based on the number of items in it. Creating the queue
with this default policy results in the queue behaving exactly as before
this change, hence existing clients can continue to use it without any
adjustments. It is worth to mention that due to the refactoring of the
code, the queue will now take one `size_t` more memory than before, in
this configuraution. I think this is acceptable given the benefits of
being able to use the queue in more advanced scenarios, e.g. where the
queue is double limited by the number of items and their collective
memory consumption.