queue: generalize limit
Allow arbitrary limits to be set. The limit is now an object
implementing the `QueueLimit` concept. The `queue` class is renamed to
`basic_queue` indicating that it is now a more generic building block.
Backward compatibility is kept by creating a `queue` class which uses a
simple size-based limit and adapts the `basic_queue` interface to that
of the former `queue` one. This adaptor has a slight overhead of
accounting the size separately (not using the `size()` of the underlying
storage).

The motivation was the need for a queue that is dual limited by element
count and total memory consumption of the elements.