queue: generalize size
Allow arbitrary types to be used as max size. This patch introduces two
concepts: `QueueSize` and `QueueItemMeasurer`. The former is the type of
the max size. The latter is the type of the functor object that is
responsible for converting queue items (of type `T`) into a size value
(of type `QueueSize`). When adding or substracting items to/from the
queue, the items's size value is obtained via the measurer and
added/substracted from the total size.
The new `Size` and `Measurer` template parameters (corresponding to the
`QueueSize` and `QueueItemMeasurer` respectively) are defaulted to
`size_t` and `counting_measurer<size_t, T>`, such that when created with
these default, `queue` will behave just like before this patch. Existing
clients can continue to use it without any adjustments.