semaphore: generalize unit
Allow arbitrary types to be used as the semaphore's unit. The type has
to fulfill the requirements of the `SemaphoreUnit` concept. The
`SemaphoreUnit` concept builds on the `QueueSize` concept, with
some additional requirements, namely the type is required to provide a
specialization for the `unit_traits<>` template. This traits object is
used to obtain the max count for the respective unit type, as well as
the equivalent of `1`. A sensible default is provided for all arithmetic
types.
A side effect of this generalization is that now both the `_count`
member and the `count` passed to the various `wait()` and `signal()`
member functions is the same type. Previously the former was signed
while the latter was unsigned. While this allowed for some the internal
count to be accidentally overflown, it prevented negative values from
being passed to these methods. This is now reversed.
The new `Unit` template parameter (corresponding to the `SemaphoreUnit`
concept) is defaulted to `ssize_t`, the semaphore remaining fully
compatible with existing code.