Make future [[nodiscard]]
After having spent (yet again) several hours chasing a bug that was
caused by a discarded future I decided it is time I did something about
this. This patch makes future [[nodiscard]], that is make the compiler
complain on each source location where a future returned from some
function is ignored. For now, this complaint is a warning, altough the
user has an option to promote it to an error, via
`--enable-error-unused-result` and `Seastar_UNUSED_RESULT_ERROR`
`configure.py` flag and `CMake` option respectively.

The intent is to allow each downstream project a transitional period to
fix up their codebase, then promote the complaint to an error to catch
any future violations on-spot. Each site where such a complaint is
suppressed (via (void)) should contain a comment with an explanation as
to why that is fine.
For now not even seastar proper checks out. I fixed all violations in
headers, as those will be complained about in each translation unit they
are included into, but there are many violations left in `.cc` files. I
added explanations to why the future is ignored, where I could deduce
it. Reviewers are encouraged to point provide explanations for the cases
where one is missing, and I will incorporate these in a next version.

Signed-off-by: Botond Dénes <bdenes@scylladb.com>