gate: introduce with_gate_wrapped(), an exception safe with_gate()
One of the most prominent usage patters of `gate` is collecting
background tasks, that are waited at a later point, perhaps by
external code. In these cases the future returned from `with_gate()` is
typically ignored, as often it is called from synchronous code. However
this leaves a glaring hole exception safety wise. Exceptions thrown by
the wrapped functions are ignored.
This patch attempts to solve this problem by providing an exception safe
`with_gate()` variant, called `with_gate_wrapped()`, that forwards any
exception raised from the wrapped function to the `gate()`, to be
re-throw from `close()`. This approach is also limited, as it can only
forward a single exception, but it will still let the continuation
waiting for the tasks to finish that there was a problem.