Per-fiber resource tracking
Controlling the concurrency of a certain internal workload based on a
maximum allowed resource consumption is a common task in server
processes. The resource consumption of the workload can sometimes be
estimated up-front with reasonable accuracy, other times this is
difficult or outright impossible. This patch attempts to address this
latter case by providing per-fiber resource tracking.

The resource tracking happens by tagging the fiber. This tag is then
propagated down the continuations chain, to all continuations initiated
from the fiber. The tag is persisted across deferring points. Tags
can be nested, that is a continuation initiated from a tagged fiber can
tag a continuation it initiates with another tag. When this nested fiber
terminates the previous tag is restored. Tags can be nested into
arbitrary depth.

The tag is an object, inheriting from the `seastar::basic_tag`
interface. The interface has two methods, `on_enter()` and `on_leave()`
which are called when the tagged fiber is started/resumed and
terminated/deferred respectively.

For now only memory tracking is implemented.

Below is a very simple use-case of resource tracking via tagging:

    struct my_tag : public seastar::basic_tag {
        uint64_t prev_memory;
        seastar::semaphore sem;

        virtual void on_enter() override {
            // memory is a public member of `basic_tag`
            prev_memory = memory;
        }
        virtual void on_leave() override {
            sem.consume_units(memory - prev_memory);
        }
    };

    auto tag = seastar::make_lw_shared<my_tag>();

    return with_semaphore(tag->sem, estimated_entry_cost, [tag] {
        return with_tag(*tag, [tag] {
            ...
        });
    });

This patch is intended as a poll of the maintainers' and peers' view on
this direction. As this patch touches quite performance sensitive parts
of seastar, I want to see if this direction is considered acceptable,
before investing more into making it complete.
The motivation for this patch is the wish to make the tracking of the
memory consumption of reads in Scylla more accurate. The plan is to tag
each read when it is admitted and update the reader concurrency
semaphore on preemption points.

TODO:
* Persist the tag in all primitives (only future::then() persists it
  now).
* Test.
* Document.