Per-fiber resource tracking
Controlling the concurrency of a certain internal workload based on a
maximum allowed resource consumption is a common task in server
processes. The resource consumption of the workload can sometimes be
estimated up-front with reasonable accuracy, other times this is
difficult or outright impossible. This patch attempts to address this
latter case by providing per-fiber resource tracking.

The resource tracking happens by tagging the fiber. This tag is then
propagated down the continuations chain, to all continuations initiated
from the fiber. The tag is persisted across deferring points. Tags
can be nested, that is a continuation initiated from a tagged fiber can
tag a continuation it initiates with another tag. When this nested fiber
terminates the previous tag is restored. Tags can be nested into
arbitrary depth.

The tag is an object, inheriting from the `seastar::basic_tag`
interface. The interface has two methods, `on_enter()` and `on_leave()`
which are called when the tagged fiber is started/resumed and
terminated/deferred respectively.

For now only memory tracking is implemented.