testing: return non-zero exit codes gracefully
Currently the way the seastar framework can propagate a non-zero exit
code is broken. We just call `exit()` in the thread hosting the
`app::run_deprecated()` function, which leads to segfault as order of
destroying components is not synchronized. The reason we are not seeing
crashes due to this is that we are currently not explicitly returning
any non-zero exit codes, so they can only come from inside the reactor.

Allow for gracefully returning non-zero exit codes by propagating them
to a member of `test_runner` and having `test_runner::finalize()` return
this. Conversely `entry_point()` is also modified to consider the exit
code returned from `test_runner::finalize()`. If both
`boost::unit_test::unit_test_main()` and `test_runner::finalize()`
returns a non-zero error code, the boost one is preferred.