testing: fail the test when there are abandoned failed futures
But only if the `--fail-on-abandoned-failed-futures` flag is set. As we
had poor visibility of these before (one had to actively look into the
test output), unconditionally failing this would cause tests to suddenly
fail en-masse.

There is a message printed unconditionally at the end, if there are any
failed futures, to improve the observability. For tests that print a lot
of messages fast, it is easy to miss the usual warning.