basic_semaphore: forward the exception factory to the expiring_fifo too
We now require that the factory be copyable, as we need to pass it to
two constructors, the semaphore's base class' (which is the exception
factory) and the `expiring_fifo`.
Maybe not a problem, in any case let's fix it, we can make it pretty
later.
Also added a test for the named timeout exception, as it was broken and
no-one noticed.