fstream: introduce safe make_file_output_stream() and make_file_data_sink()
Currently these functions are unsafe in that if they fail to construct
the output_stream or the data_sink respectively the `file` instance
passed in will be left unclosed, triggering an assertion failure in the
`~file()`. As these classes take ownership of the file after being
constructed, it makes sense for them to take care of the `file` object
from the moment it is passed in.

For this to be safe the functions have to return a future<>, so that
in the event of a failure callers can wait for the file to be closed.
Since these functions are part of seastar's public API we can't just
change their return type. To allow for transition to the new version,
the safe version is put in a `safe` namespace, while the old versions
are put in an inline `unsafe` namespace, and a deprecation warning is
added to them. Old code will seamlessly keep using the old version by
default, the compiler warning about the deprecation. Code can be
transitioned to the `safe::` versions. After a while the old versions
will be removed and the `safe` namespace will be made inline, allowing
for new code to use them with or without `safe::` prefix.