fstream: introduce safe make_file_output_stream() and make_file_data_sink()
Currently these functions are unsafe in that if they fail to construct
the output_stream or the data_sink respectively the `file` instance
passed in will be left unclosed, triggering an assertion failure in the
`~file()`. As these classes take ownership of the file after being
constructed, it makes sense for them to take care of the `file` object
from the moment it is passed in.

For this to be safe the functions have to return a future<>, so that
in the event of a failure callers can wait for the file to be closed.
The new safe variants are added to api_v3 namespace (also created
in this patch).