memory: do_dump_memory_diagnostics(): print to log_buf
Use the writer `logger::log()` API to format the message into the log
buffer directly. This should ensure that formatting itself doesn't
allocate at all -- which is important given that we are logging right
after a failed allocation. It also produces a single log message,
contrary to the previous state where each line was a separate log
message, possibly interwoven with other log lines.
To make this a little more readable the report is adjusted -- the first
log line just explains that a detailed report will follow, the report
proper starts on the next new line. Formatting of the report itself is
not changed.