memory: do_dump_memory_diagnostics(): print to std::ostream&
Instead of logging each line separately, take an `std::ostream&` and
write into that. This allows abstracting away the final destination of
the report as well as collecting all the output into a single
`std::ostringstream` and logging it as a single message, helping to
ensure the report is not interwoven with other log lines.
For now we just to the latter -- collect all the output into a single
`std::ostringstream` and log it as a single message. To make this a
little more readable the report is adjusted -- the first log line just
explains that a detailed report will follow, the report proper starts on
the next new line. Formatting of the report itself is not changed.