memory: do_dump_memory_diagnostics(): print numbers in a compact and human readable way
Convert sizes to the closest multiple of 2^(N*10) and use the
appropriate suffix (B, K, M, G, T, E). E.g. 2K instead of 2048.
Convert numbers to the closest multiple of 10^N and use the appropriate
suffix (k, m, b, t), E.g. 34k instead of 341253.
Convert percentage calculated as double to unsigned.

This makes the report more compact but more human readable as well at
the cost of losing some information. But as this report is only supposed
to be the first step of an investigation, this is acceptable.