memory: introduce internal::emergency_allocator
A special purpose allocator for cases where one really wants allocations to
succeed even though memory is known to be very low. One such example is
logging the memory diagnostics on allocation failures.

The allocator works by trying to allocate objects in higher size pools
when it fails in the size appropriate one. As such it doesn't support
the free() path where size is also provided. For this reason it
implements the `std::allocator` interface, where we have control over
the free path too.

The allocator cannot do miracles, it is not backed by emergency
reserves, so when memory is really all gone it will fail just the same.
Furthermore it is quite wasteful as in the extreme case it can allocate
a 2K slot for an 8B object. So it is intended to be used for emergency
situations and for short-lived allocations.