log: introduce emergency_log()
This log path uses the emergency allocator
`seastar::memory::internal::emergency_allocator` to allocate the
buffers needed for the logging. This is an internal API intended to be
used in emergency situations, like logging a memory diagnostics report
upon allocation failure -- the use-case which motivated writing it.

The catch is that this only works in C++20, as `std::stringstream`
doesn't have an `Allocator` template param in prior standard versions.