log: logger::do_log(): reduce allocations even more
Use the recently introduced log_buf as the buffer the entire log message
is formatted into. This is now backed by a static 8KB buffer that is
reused between all calls of `logger::do_log()`. Allocations are only
needed if the message size is larger than this static buffer.
To allow callers to format their messages directly into this buffer a
new low level entry point is introduced. This passes a reference to a
writer object. This object will provided direct access to the underlying
log buffer. Using this patch, one can format the entire log message
completely without allocations, as long as the message size stays
below said 8KB.
Note that `do_log()` is still not completely noexcept as the sinks
themselves (`std::cerr` and `syslog()`) might still allocate.