memory: allow configuring when to dump memory diagnostics report
This can be configured based on the kind of the allocation failure.
Currently there are three options:
* none
* critical
* all

Critical allocations can be marked as such (for an entire scope) with
`scoped_critical_alloc_section`, provided by this patch.

Note that to preserve pre-existing behaviour, when the seastar_logger is
set to log level >= debug, a report is dumped regardless of this
configuration.