reactor: add counter for broken promises
Broken promises always signal bugs. Add a counter to make them
observable. The best place to increase the counter is
`set_to_broken_promise()`. To give it access to the counter in the
reactor we make it a friend of the latter, and to be able to do that we
have to un-static it and expose it in the `future.hh` header.