testing: log number of broken promises at the end
To make them more visible. They might not be logged at all when actually
happening and even if they are it is just lost in the noise.
We don't add a test failing facility similar to
--fail-on-abandoned-failed-futures, as we want to move away from flags
like that, instead adding a flag to abort the application when the
respective error happens, which makes for a much better debugging
experience.