futures: add set_abort_on_broken_promises()
The best way to diagnose a broken promise error is to get a coredump
with the stack that produced it. This patch adds the infrastructure
to enable that, a flag which when set, triggers an abort on broken
promise errors. For now the flag can only be set from code.