reactor: introduce single command-line flag for enabling abort on various errors
We have a growing number of abort-on-something flags that can be set
from the code to abort when the respective class of errors happen,
making debugging easier. Some of these are exposed via command line
flags as well, but not all, one example is
`--abort-on-seastar-bad-alloc`. We want to add yet another of these
flags and expose it via the command line as well, but the current method
is not scalable, it will quickly result in an explosion of command-line
flags (we have a lot already). To solve this this patch adds a single
flag to cover all the above mentioned abort-on-something cases. This
flag accepts multiple values, allowing the user to set any subset of the
available flags, or all of them with the keyword "all".