app-template: introduce tool mode
Some seastar-based projects might want to add tools to their arsenal
that are based on the main application's code-base. Currently seastar
makes this quite awkward as its default settings cater to the highly
performance-sensitive applications it was designed for.
This patch provides a solution to this by adding a `mode` parameter to
`app_template::config`. Mode has two values:
* app - which is identical to the previous status-quo;
* tool - which is a mode catering to the above mentioned tools;

In tool mode, seastar hides all its internal options from `--help`
output (except logging options), and it overrides the default values of
certain options to make the tool act less like it owns the machine it
runs on. In particular by default a single CPU will be used, along with
100MB memory and all the various tuneables are relaxed as much as
possible.