app-template: introduce seastar_options
This is the top-level option struct that will hold the configuration for
all of seastar. Currently unused, we will add content gradually in the
following patches. After all of seastar can be configured through it we
will add a new app_template constructor accepting a `seastar_options`
instance, allowing seastar to be (pre)configured by C++ code.