log: make timestamp style a logger-scoped global
To simplify further patching. This brings timestamp style into the fold:
it is now an atomic, private global on the logger class, updated via a
setter method. Setting the timestamp style on the logger also sets the
`print_timestamp` function pointer. Also update the default of said
pointer to match that set in the CLI description.