log: add log_cli::apply_logging_settings
A variant of the existing `log::apply_logging_settings()` which takes
raw boost program options variable map. This allows changing only the
settings that are actually set to a non-default value, preventing
blanket setting everything to non-default. This allows the conflict
resolution to come into play. An alternative source of settings (e.g. a
config-file) can now set settings safely without overriding the CLI
provided ones (if that is desired).