log: introduce conflict resolution infrastructure
Will be used by the next patches to allow resolving conflicting settings
coming from different sources in a controlled manner.
For this we need to know if we have a conflict in the first place. We
define a conflict when the value of an already changed setting is
changed again. Normally this is not a concern, but with logging we want
to prioritize settings values coming from different sources.
The infrastructure introduced here allows two ways to handle this:
* previous value has priority
* new value has priority
If the value is still on a default setting (which is also tracked now)
any change is allowed.