utils/program_options: deprecate string_map
This class doesn't have any internal uses since 8810950b. Since there
might be external users (we know of at least one such user: scylla), we
cannot just remove it. We deprecate it and will remove it after some
grace period.
Even though this class is still a useful CLI utility, seastar is not in
the business of providing CLI utilities so it is best to remove it.